+++ src/main/java/com/photosend/photosendserver01/domains/user/domain/ProductRepository.java
@@ -2,6 +2,6 @@ package com.photosend.photosendserver01.domains.user.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
-    ProductEntity findByPidAndUserWechatUid(Integer pid, String wechatUid);
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
+    ProductEntity findByPidAndUserWechatUid(Long pid, String wechatUid);
 }
